prettified arbitrum

+++ contracts/aggregator/arbitrum/flashloan/interfaces.sol
@@ -25,12 +25,23 @@ interface ListInterface {
 
 interface TokenInterface {
     function approve(address, uint256) external;
-    function transfer(address, uint) external;
-    function transferFrom(address, address, uint) external;
+
+    function transfer(address, uint256) external;
+
+    function transferFrom(
+        address,
+        address,
+        uint256
+    ) external;
+
     function deposit() external payable;
-    function withdraw(uint) external;
+
+    function withdraw(uint256) external;
+
     function balanceOf(address) external view returns (uint256);
+
     function decimals() external view returns (uint256);
+
     function totalSupply() external view returns (uint256);
 }
 
@@ -45,5 +56,9 @@ interface IBalancerLending {
         uint256[] memory amounts,
         bytes memory userData
     ) external;
-    function getProtocolFeesCollector() external view returns (ProtocolFeesCollector);
+
+    function getProtocolFeesCollector()
+        external
+        view
+        returns (ProtocolFeesCollector);
 }
